+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -104,7 +104,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kfb0s059b050mszyqwh5d88z"
+        "conv_01kfb68tc0nrng0vbgrdyaf9sn"
     )
     r"""Unique conversation identifier with `conv_` prefix."""