+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFB0S0Z9C54XF6X3QVEZK5EC"
+        "01KFB68V3ZKDDGHEPWWAVMDQ3D"
     )
     r"""The unique identifier of the data source"""