+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -154,7 +154,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-19T11:38:09.062Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-19T13:14:11.941Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")