+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFB0S0ZDPA99XJ1RSMP4CC9T"
+        "01KFB68V48TJTT2BVYH1K282FB"
     )
     r"""The unique identifier of the data source"""